api: fold music curation helpers into owning pair and repo
