+++ swift/api/Sources/Api/PairQL/Dashboard/Pairs/ApproveMusicArtist_v2.swift
@@ -60,9 +60,9 @@ extension ApproveMusicArtist_v2: Resolver {
         resolvedAt: now,
         in: db,
       )
-      let snapshot = try await publishMusicPolicy(
+      let snapshot = try await Music.LibrarySnapshotRepository.publish(
         childId: child.id,
-        changed: changed,
+        policyChanged: changed,
         generatedAt: now,
         in: db,
       )
