+++ swift/api/Sources/Api/PairQL/Dashboard/Pairs/ApproveMusicTrack.swift
@@ -32,9 +32,9 @@ extension ApproveMusicTrack: Resolver {
         resolvedAt: now,
         in: db,
       )
-      let snapshot = try await publishMusicPolicy(
+      let snapshot = try await Music.LibrarySnapshotRepository.publish(
         childId: child.id,
-        changed: changed,
+        policyChanged: changed,
         generatedAt: now,
         in: db,
       )
