+++ swift/api/Sources/Api/PairQL/Dashboard/Pairs/SaveMusicAlbumCuration.swift
@@ -73,9 +73,9 @@ extension SaveMusicAlbumCuration: Resolver {
         resolvedAt: now,
         in: db,
       )
-      let snapshot = try await publishMusicPolicy(
+      let snapshot = try await Music.LibrarySnapshotRepository.publish(
         childId: child.id,
-        changed: changed,
+        policyChanged: changed,
         generatedAt: now,
         in: db,
       )
